+++ src/outputFormat.ts
@@ -1,0 +1,6 @@
+'use strict';
+
+export enum OutputFormat {
+  PlainText = 'plaintext',
+  Json = 'json'
+}